requirements.txt
fastapi==0.95.0
pydantic==1.10.7
SQLAlchemy==2.0.10
uvicorn[standard]
gunicorn==20.1.0
aiohttp==3.8.4
python-jose==3.3.0
passlib==1.7.4
GeoAlchemy2==0.13.2